Fix logic for suggested id for new site.
This would take the default "Plone" plus the count of existing sites.
Usually this works fine, first suggesting Plone, then Plone1, Plone2, etc.
But if you have Plone and Plone2 as existing site ids, it would suggest Plone2, which gives an error because it is already taken.

The fix is: keep this logic, but if the id is already taken, keep increasing the id until it is free.

This fixes https://github.com/plone/plone.distribution/issues/97